requirements.txt
python-dotenv~=1.0.0
asyncio~=3.4.3
python-telegram-bot
functions_framework
firebase-admin
